+++ srcs/ft_print_before.c
@@ -22,9 +22,9 @@ static int	print_before_minus(t_flags *flag, int c_p, char *s, char c)
 		&& flag->prec != -1 && s[0] != '\0')
 	{
 		if (c == 'o')
-			c_p += ft_putcx('0', flag->prec - c_p - ft_strlen(s));
+			c_p += ft_putcx('0', flag->prec - c_p - (int)ft_strlen(s));
 		else
-			c_p += ft_putcx(' ', flag->prec - c_p - ft_strlen(s));
+			c_p += ft_putcx(' ', flag->prec - c_p - (int)ft_strlen(s));
 	}
 	c_p += ft_putnchar(s, ft_strlen(s));
 	if (flag->zero == 1)
@@ -53,8 +53,8 @@ static int	print_before_plus(t_flags *flag, int c_p, char *s, char c)
 	if (flag->plus == 1)
 		c_p = is_plus(flag, c_p, &s, c);
 	if (flag->width - c_p > (int)ft_strlen(s) && flag->zero != 1)
-		c_p += ft_putnchar(" ", flag->width - (int)ft_strlen(s));
-	c_p += ft_putcx('0', flag->width - ft_strlen(s) - c_p);
+		c_p += ft_putnchar(" ", (size_t)flag->width - ft_strlen(s));
+	c_p += ft_putcx('0', flag->width - (int)ft_strlen(s) - c_p);
 	if (c == '\0')
 		print_c('\0');
 	else
@@ -85,13 +85,13 @@ static int	just_print(t_flags *flag, int c_p, char *s, char c)
 		c_p += 2;
 	if ((c == 'o' || c == 'u' || c == 'x' || c == 'X')
 		&& (flag->prec > (int)ft_strlen(s)))
-		c_p += flag->prec - ft_strlen(s) - flag->plus;
+		c_p += flag->prec - (int)ft_strlen(s) - flag->plus;
 	if (c == '\0')
 		c_p += ft_putcx(' ', flag->width - 1 - flag->plus - c_p);
 	else
-		c_p += ft_putcx(' ', flag->width - ft_strlen(s) - flag->plus - c_p);
+		c_p += ft_putcx(' ', flag->width - (int)ft_strlen(s) - flag->plus - c_p);
 	if (c == 'o' || c == 'u' || c == 'x' || c == 'X')
-		c_p += ft_putcx('0', flag->prec - ft_strlen(s) - flag->plus);
+		c_p += ft_putcx('0', flag->prec - (int)ft_strlen(s) - flag->plus);
 	if (flag->sharp == 1)
 		c_p = is_sharp(c_p, c);
 	if (flag->plus == 1)
